==> code/3_SocioEnviroPredictors.Rmd <==
---
title: "Social and Environmental Predictors of Microbiome Age"
author: "Mauna Dasari"
output:
  html_document:
    number_sections: no
    theme: paper
    toc: yes
    toc_depth: 3
    toc_float: yes
  pdf_document:
    toc: yes
    toc_depth: '3'
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = F,warning=F,message=F,cache=2)
library(feather);library(tidyverse);library(kableExtra);library(gridExtra);

library(lme4);library(lmerTest);library(mgcv);library(survival);library(survminer);
metadata_all<-readRDS("../data/metadata_full.rds") 
metadata_all$collect_month<-as.factor(metadata_all$collect_month)

H.PredLife_raw<-readRDS("../data/HetGPR.PredictedAgeData.rds")

PredLifeCov<-H.PredLife_raw %>% 
  select(-c(testID,delta_abs)) %>% 
  left_join(metadata_all, by=c("DADA_id","sname")) 
#saveRDS(PredLifeCov,"../data/PredictedAge_EnviroCov.rds")
```

# What predicts delta age? {.tabset}
$\Delta$age = $age_{microbial}$ - $age_{chronological}$, so if $\Delta$age is negative, that means $age_{chronological}$ > $age_{microbial}$ and the animal is microbially young for age. 


## Males {.tabset}
### Sources of Adversity {.tabset}
```{r}
PredLife_M.B<-PredLifeCov %>% filter(sex=="M")
model1<-lmer(delta~age_chrono+
               avg_month_maxtemp+
               sum_month_rain+
               season+
               maternal_loss+
               sibling+
               drought+
               highGrpSize+
               lowMatSCI+
               lowMatRank+
               ordrank+
               (1|sname)+
               (1|collect_grp)+
               (1|hydroyear),
             data=PredLife_M.B)
summary(model1)
```

### Cumulative Adversity {.tabset}
```{r}
PredLife_M.B<-PredLifeCov %>% filter(sex=="M")
model<-lmer(delta~age_chrono+
               avg_month_maxtemp+
               sum_month_rain+
               season+
               cumulative_adversity+
               ordrank+
               (1|sname)+
               (1|collect_grp)+
               (1|hydroyear),
             data=PredLife_M.B)
summary(model)
```

## Females {.tabset}
### Sources of Adversity {.tabset}
```{r}
PredLife_F.B<-PredLifeCov %>% filter(sex=="F")
model1<-lmer(delta~age_chrono+
               avg_month_maxtemp+
               sum_month_rain+
               season+
               maternal_loss+
               sibling+
               drought+
               highGrpSize+
               lowMatSCI+
               lowMatRank+
               proprank+
               adf_collect+
               (1|sname)+
               (1|collect_grp)+
               (1|hydroyear),
             data=PredLife_F.B) 
summary(model1)
```

### Cumulative Adversity {.tabset}
```{r}
PredLife_F.B<-PredLifeCov %>% filter(sex=="F")

model<-lmer(delta~age_chrono+
               avg_month_maxtemp+
               sum_month_rain+
               season+
               cumulative_adversity+
               proprank+
               adf_collect+
               (1|sname)+
               (1|collect_grp)+
               (1|hydroyear),
             data=PredLife_F.B) 
summary(model)
```
