==> code/4_MilestoneAnalysis.Rmd <==
---
title: "Estimating Onset of Maturational Milestones and/or Survival"
author: "Mauna Dasari"
output:
  html_document:
    number_sections: no
    theme: paper
    toc: yes
    toc_depth: 3
    toc_float: yes
  pdf_document:
    toc: yes
    toc_depth: '3'
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = F,warning=F,message=F,cache=2)
library(feather);library(tidyverse);library(kableExtra);library(gridExtra);
library(lme4);library(lmerTest);library(mgcv);library(survival);library(survminer);


H.PredLife_raw<-readRDS("../data/HetGPR.PredictedAgeData.rds")

metadata_all<-readRDS("../data/metadata_full.rds") %>% 
  left_join(H.PredLife_raw,by=c("DADA_id","sname","age_chrono","sex"))

count<-metadata_all %>% 
  group_by(sname) %>% 
  mutate(count_life=n()) %>% 
  ungroup() 

females<-metadata_all %>% 
  left_join(count,by="sname") %>% 
  filter(sex=="F"&count_life>3)
males<-metadata_all %>%  
  left_join(count,by="sname") %>% 
  filter(sex=="M"$count_life>3)

females$resid_life<-resid(lmer(age_mb~age_chrono+ #AIC 37365.56
                                      avg_month_maxtemp+
                                      sum_month_rain+
                                      season+
                                      (1|collect_grp)+
                                      (1|hydroyear),
                                    data=females,
                                    control=lmerControl(optCtrl=list(maxfun=2000))))
males$resid_life<-resid(lmer(age_mb~age_chrono+ #AIC 37365.56
                                      avg_month_maxtemp+
                                      sum_month_rain+
                                      season+
                                      (1|collect_grp)+
                                      (1|hydroyear),
                                    data=males,
                                    control=lmerControl(optCtrl=list(maxfun=2000))))

```

# Introduction
## Does microbiome age predict maturation and other milestones? 
*originally part of the HetGPR_clean file, but split out for ease of access on 11/24/2020.*

Or, what are the consequences of variation in microbial age? Here, I'll take the residuals of $age_{microbial}$~$age_{chronological}$+$seasonal covariates$ (see below) as microbial acceleration and the slope as 'pace of aging' to see how predictive these markers are of the timing of maturational milestones and death.

***

**Residuals**

First, we compared whether it made more sense to account for covariates or not. Since it didn't seem to make a difference in residual spread and because the AICs of the covariate included models are lower, I went with the residuals corrected for covariates.

Additionally, these residuals only included the animals that either matured *within* the sampling period or did not mature and had a status of 0, 2, or 3 (to be censored).

# Female Rank Attainment
## Prior to Milestone
### Data 
```{r}
PriorLadyRank<-females %>% 
  mutate(age_rankedf=as.numeric((rankedf-birth)/365.25)) %>% 
  filter(case_when(!is.na(rankedf)~rankedf>min(metadata_all$collection_date),
                   is.na(rankedf)~age_chrono<max(age_rankedf,na.rm=T)&
                     (status==0|status==2|status==3))) %>% 
  filter(age_chrono<max(age_rankedf,na.rm=T))
PriorLadyRank$resid_cov<-resid(lmer(age_mb~age_chrono+ 
                                      avg_month_maxtemp+
                                      sum_month_rain+
                                      season+
                                      (1|collect_grp)+
                                      (1|hydroyear),
                      data=PriorLadyRank,
                      control=lmerControl(optCtrl=list(maxfun=2000))))

Summary_PriorLadyRank<-PriorLadyRank %>%
  group_by(sname) %>% 
  mutate(resid_avg_priorrankedf=mean(resid_cov),
         rain_avg_priorrankedf=mean(rain_yr,na.rm=T),
         matsis_avg_priorrankedf=mean(matsisters_yr,na.rm=T),
         adf_avg_priorrankedf=mean(avg_adf_yr,na.rm=T),
         age_avg_priorrankedf=mean(age_chrono)) %>%  
  ungroup() %>% 
  group_by(sname,sex,status,birth,statdate,rankedf,resid_avg_priorrankedf,age_rankedf,rnktype_focal,matsis_avg_priorrankedf,mompresent_rankedf,adf_avg_priorrankedf,rain_avg_priorrankedf,hybridscore,lowMatRank,age_avg_priorrankedf,genome_wide_anubis_ancestry) %>% 
  summarise() %>% 
  ungroup() %>% 
  mutate(time=case_when(is.na(rankedf)~(statdate-birth)/365.25,
                       !is.na(rankedf)~(rankedf-birth)/365.25),
         event=case_when(is.na(rankedf)~0,
                         !is.na(rankedf)&rnktype_focal=="ADF"~0,
                         !is.na(rankedf)&rnktype_focal=="ALF"~1),
         type_cov_accel=case_when(resid_avg_priorrankedf>
                                    median(resid_avg_priorrankedf)~"old",
                                  resid_avg_priorrankedf<
                                    median(resid_avg_priorrankedf)~"young",
                                  resid_avg_priorrankedf==
                                    median(resid_avg_priorrankedf)~"young")) %>% 
  ungroup() 
#saveRDS(Summary_PriorLadyRank,"../out/DataForFigures/rankattainment_females_prior_linear_2022.rds")
```


### Plots
```{r}
survfit.cov.PriorLadyRank<-survfit(Surv(time,event)~type_cov_accel,
                          data=Summary_PriorLadyRank)
p.coxph.cov<-ggsurvplot(survfit.cov.PriorLadyRank, 
                 data = Summary_PriorLadyRank, 
                 fun="event",
                 title = "Age Acceleration Curves Prior to Female Rank Attainment",
                 ylab = "Probability of Female Rank Attainment",
                 xlab = "Time in Years",
                 break.x.by = 1,
                 xlim=c(0,7),
                 conf.int = T,
                 pval = F, pval.method = F,    ### Add p-value &  method name
                 surv.median.line = "hv",         ### Add median survival lines
                 legend.title = "Age Acceleration",  ### Change legend titles
                 legend.labs = c("Old", "Young"), ### Change legend labels
                 palette = "nejm",         ### Use JCO journal color palette
                 risk.table = F,                  ### Add No at risk table
                 cumevents = F,                   ### Add cumulative No of events table
                 tables.height = 0.15,            ### Specify tables height
                 tables.theme = theme_cleantable(),### Clean theme for tables
                 tables.y.text = FALSE            ### Hide tables y axis text
                 )
p.coxph.cov$plot<-p.coxph.cov$plot+geom_vline(xintercept = 2.24, linetype = "solid", color="red")
p.coxph.cov
```

### Models {.tabset}
```{r}
coxph.LadyRank.cov_original<-coxph(Surv(time,event)~resid_avg_priorrankedf
                        #  +slope_rankedf
                          #+age_avg_priorrankedf
                          +mompresent_rankedf
                          +matsis_avg_priorrankedf
                          +lowMatRank
                          +adf_avg_priorrankedf
                          +rain_avg_priorrankedf
                          +genome_wide_anubis_ancestry
                          ,data=Summary_PriorLadyRank)
sink("model.txt")
summary(coxph.LadyRank.cov_original);cox.zph(coxph.LadyRank.cov_original)
sink()
ggforest(coxph.LadyRank.cov_original,data=Summary_PriorLadyRank,
         main="Covariates Prior to Female Rank Attainment")

coxph.LadyRank.cov_ageexplicit<-coxph(Surv(time,event)~accel_avg_priorrankedf
                          #+slope_rankedf
                          +age_avg_priorrankedf
                          +mompresent_rankedf
                          +matsis_avg_priorrankedf
                          +lowMatRank
                          +adf_avg_priorrankedf
                          +rain_avg_priorrankedf
                          +genome_wide_anubis_ancestry
                          ,data=Summary_PriorLadyRank)
summary(coxph.LadyRank.cov_original);cox.zph(coxph.LadyRank.cov_original)
ggforest(coxph.LadyRank.cov_original,data=Summary_PriorLadyRank,
         main="Covariates Prior to Female Rank Attainment")
```

# Menarche {.tabset}
## Prior to Milestone {.tabset .tabset-pills}
### Data 
```{r}
PriorMenarche<-females %>%  
  filter(case_when(!is.na(matured)~matured>min(metadata_all$collection_date)&mstatus=="O",
                   is.na(matured)~(status==0|status==2|status==3))) %>% 
  filter(age_chrono<max(age_matured,na.rm=T))

PriorMenarche$resid_cov<-resid(lmer(age_mb~age_chrono+ 
                                      avg_month_maxtemp+
                                      sum_month_rain+
                                      season+
                                      (1|collect_grp)+
                                      (1|hydroyear),
                      data=PriorMenarche,
                      control=lmerControl(optCtrl=list(maxfun=2000))))

Summary_PriorMenarche<-PriorMenarche %>%
  group_by(sname) %>% 
  mutate(resid_avg_priormenarche=mean(resid_cov),
         rain_avg_priormenarche=mean(rain_yr,na.rm=T),
         matsis_avg_priormenarche=mean(matsisters_yr,na.rm=T),
         adf_avg_priormenarche=mean(avg_adf_yr,na.rm=T),
         age_avg_priormenarche=mean(age_chrono)) %>%  
  ungroup() %>% 
  group_by(sname,sex,status,birth,statdate,matured,resid_avg_priormenarche,mstatus,age_matured,matsis_avg_priormenarche,mompresent_mat,adf_avg_priormenarche,rain_avg_priormenarche,hybridscore,lowMatRank,age_avg_priormenarche,genome_wide_anubis_ancestry) %>% 
  summarise() %>% 
  ungroup() 

Summary_PriorMenarche<-Summary_PriorMenarche %>% 
  mutate(time=case_when(is.na(matured)~(statdate-birth)/365.25,
                       !is.na(matured)~(matured-birth)/365.25),
         event=case_when(is.na(matured)~0,
                         !is.na(matured)~1),
         type_cov_accel=case_when(resid_avg_priormenarche>
                                    median(resid_avg_priormenarche)~"old",
                                  resid_avg_priormenarche<
                                    median(resid_avg_priormenarche)~"young",
                                  resid_avg_priormenarche==
                                    median(resid_avg_priormenarche)~"young")) %>% 
  ungroup() %>% drop_na()
#saveRDS(Summary_PriorMenarche,"../data/menarche_prior.rds")
```

### Plots
```{r}
survfit.cov.PriorMenarche<-survfit(Surv(time,event)~type_cov_accel,
                          data=Summary_PriorMenarche)
p.coxph.cov<-ggsurvplot(survfit.cov.PriorMenarche, 
                 data = Summary_PriorMenarche, 
                 fun="event",
                 title = "Age Acceleration Curves Prior to Menarche",
                 ylab = "Probability of Menarche",
                 xlab = "Time in Years",
                 break.x.by = 0.5,
                 xlim=c(2.5,6.5),
                 conf.int = T,
                 pval = TRUE, pval.method = F,    ### Add p-value &  method name
                 surv.median.line = "hv",         ### Add median survival lines
                 legend.title = "Age Acceleration",  ### Change legend titles
                 legend.labs = c("Old", "Young"), ### Change legend labels
                 palette = "nejm",         ### Use JCO journal color palette
                 risk.table = F,                  ### Add No at risk table
                 cumevents = F,                   ### Add cumulative No of events table
                 tables.height = 0.15,            ### Specify tables height
                 tables.theme = theme_cleantable(),### Clean theme for tables
                 tables.y.text = FALSE            ### Hide tables y axis text
                 )
p.coxph.cov$plot<-p.coxph.cov$plot+geom_vline(xintercept = 4.51, linetype = "solid", color="red")
p.coxph.cov

```

### Models {.tabset}
```{r}
coxph.menarche.cov<-coxph(Surv(time,event)~resid_avg_priormenarche
                          #+type_cov_pace
                          #+age_avg_priormenarche
                          +mompresent_mat
                          +matsis_avg_priormenarche
                          +lowMatRank
                          +adf_avg_priormenarche
                          +rain_avg_priormenarche
                          +genome_wide_anubis_ancestry
                          ,data=Summary_PriorMenarche)
sink("model.txt")
AIC(coxph.menarche.cov);summary(coxph.menarche.cov); 
cox.zph(coxph.menarche.cov)
sink()
ggforest(coxph.menarche.cov,data=Summary_PriorMenarche,
         main="Covariates Prior to Menarche")

coxph.menarche.cov<-coxph(Surv(time,event)~type_cov_accel
                          #+type_cov_pace
                          #+age_avg_priormenarche
                          +mompresent_mat
                          +matsis_avg_priormenarche
                          +lowMatRank
                          +adf_avg_priormenarche
                          +rain_avg_priormenarche
                          +genome_wide_anubis_ancestry
                          ,data=Summary_PriorMenarche)
AIC(coxph.menarche.cov);summary(coxph.menarche.cov); 
cox.zph(coxph.menarche.cov)
ggforest(coxph.menarche.cov,data=Summary_PriorMenarche,
         main="Covariates Prior to Menarche")

#fit0<-survfit(Surv(age_death_or_censored,censored)~allsame,data=eli_fq)
#fit0 #to get median survival times
```

# First Live Birth {.tabset .tabset-pills}
## Prior to Milestone {.tabset}
### Data
```{r}
Priorfirstlivebirth<-females %>%  
  filter(case_when(!is.na(date_firstlivebirth)~
                     date_firstlivebirth>min(metadata_all$collection_date),
                   is.na(date_firstlivebirth)~status==0|status>1)) %>% 
  filter(age_chrono<max(age_firstlivebirth,na.rm=T)) 

Priorfirstlivebirth$resid_cov<-resid(lmer(age_mb~age_chrono+ 
                                      avg_month_maxtemp+
                                      sum_month_rain+
                                      season+
                                      (1|collect_grp)+
                                      (1|hydroyear),
                                    data=Priorfirstlivebirth,
                                    control=lmerControl(optCtrl=list(maxfun=2000))))

Summary_Priorfirstlivebirth<-Priorfirstlivebirth %>% 
  group_by(sname) %>% 
  mutate(resid_avg_prior=mean(resid_cov),
         adf_avg_prior=mean(avg_adf_yr,na.rm=T),
         matsis_avg_prior=mean(matsisters_yr,na.rm=T),
         rain_avg_prior=mean(rain_yr,na.rm=T),
         age_avg_prior=mean(age_chrono)) %>% 
  ungroup() %>% 
  group_by(sname,sex,birth,statdate,resid_avg_prior,age_firstlivebirth,date_firstlivebirth,mompresent_lb,hybridscore,adf_avg_prior,matsis_avg_prior,rain_avg_prior,lowMatRank,age_avg_prior,genome_wide_anubis_ancestry) %>% 
  summarise() %>% 
  ungroup() %>% 
  mutate(time=case_when(is.na(date_firstlivebirth)~(statdate-birth)/365.25,
                       !is.na(date_firstlivebirth)~(date_firstlivebirth-birth)/365.25),
         event=case_when(is.na(date_firstlivebirth)~0,
                         !is.na(date_firstlivebirth)~1),
         type_cov_accel=case_when(resid_avg_prior>median(resid_avg_prior)~"old",
                            resid_avg_prior<median(resid_avg_prior)~"young",
                            resid_avg_prior==median(resid_avg_prior)~"old")) %>% 
  ungroup()
```

### Plots
```{r}
survfit.cov.Priorfirstlivebirth<-survfit(Surv(time,event)~type_cov_accel,
                          data=Summary_Priorfirstlivebirth)
p.coxph.cov<-ggsurvplot(survfit.cov.Priorfirstlivebirth, 
                 data = Summary_Priorfirstlivebirth, 
                 fun="event",
                 title = "First LB Curves for Age Accel Prior to Milestone",
                 ylab = "Probability of First LB",
                 xlab = "Time in Years",
                 break.x.by = 0.5,
                 xlim=c(4.5,10),
                 conf.int = T,
                 pval = TRUE, pval.method = F,    ### Add p-value &  method name
                 surv.median.line = "hv",         ### Add median survival lines
                 legend.title = "Age Acceleration",  ### Change legend titles
                 legend.labs = c("Old", "Young"), ### Change legend labels
                 palette = "nejm",         ### Use JCO journal color palette
                 risk.table = F,                  ### Add No at risk table
                 cumevents = F,                   ### Add cumulative No of events table
                 tables.height = 0.15,            ### Specify tables height
                 tables.theme = theme_cleantable(),### Clean theme for tables
                 tables.y.text = FALSE            ### Hide tables y axis text
                 )
p.coxph.cov$plot<-p.coxph.cov$plot+geom_vline(xintercept = 5.97, linetype = "solid", color="red")
p.coxph.cov
```

### Models {.tabset}
```{r}
coxph.firstlivebirth.cov<-coxph(Surv(time,event)~resid_avg_prior
                                #+slope_covfirstlb
                               # +age_avg_prior
                                +mompresent_lb
                                +matsis_avg_prior
                                +lowMatRank
                                +adf_avg_prior
                                +rain_avg_prior
                                +genome_wide_anubis_ancestry
                           ,data=Summary_Priorfirstlivebirth)
sink("model.txt")
summary(coxph.firstlivebirth.cov);cox.zph(coxph.firstlivebirth.cov)
sink()
ggforest(coxph.firstlivebirth.cov,data=Summary_Priorfirstlivebirth,
         main="Covariates for Age Accel Prior to First LB")
```

# Testicular Enlargement {.tabset .tabset-pills}
## Prior to Milestone {.tabset}
### Data
```{r}
PriorTE<-males %>%  
  filter(case_when(!is.na(matured)~matured>min(metadata_all$collection_date)&mstatus=="O",
                   is.na(matured)~status==0|status==2|status==3)) %>% 
  filter(age_chrono<=max(matured,na.rm=T)) %>% 
  drop_na(hydroyear)

PriorTE$resid_cov<-resid(lmer(age_mb~age_chrono+
                                avg_month_maxtemp+
                                sum_month_rain+
                                season+
                                (1|collect_grp)+
                                (1|hydroyear),
                              data=PriorTE,
                              control=lmerControl(optCtrl=list(maxfun=2000))))

Summary_PriorTE<-PriorTE %>% 
  left_join(slope_M,by="sname") %>% 
  group_by(sname) %>% 
  mutate(resid_avg_TE=mean(resid_cov,na.rm=T),
         rain_avg_TE=mean(rain_yr,na.rm=T),
         matsis_avg_TE=mean(matsisters_yr,na.rm=T),
         excess_cfemales_avg_TE=mean(excess_cfemales_yr,na.rm=T),
         age_avg_TE=mean(age_chrono,na.rm=T)) %>% 
  ungroup() %>% 
  group_by(sname,sex,birth,statdate,matured,resid_avg_TE,mstatus,age_matured,matsis_avg_TE,hybridscore,excess_cfemales_avg_TE,rain_avg_TE,mompresent_mat,lowMatRank,age_avg_TE,genome_wide_anubis_ancestry) %>% 
  summarise() %>% 
  ungroup() %>% 
  mutate(time=case_when(is.na(matured)~(statdate-birth)/365.25,
                       !is.na(matured)~(matured-birth)/365.25)) %>% 
  mutate(event=case_when(is.na(matured)~0,
                         !is.na(matured)~1)) %>% 
  mutate(type_cov_accel=case_when(resid_avg_TE>median(resid_avg_TE)~"old",
                            resid_avg_TE<median(resid_avg_TE)~"young",
                            resid_avg_TE==median(resid_avg_TE)~"young"))
```

### Plots
```{r}
survfit.cov.PriorTE<-survfit(Surv(time,event)~type_cov_accel,
                          data=Summary_PriorTE)
p.coxph.cov<-ggsurvplot(survfit.cov.PriorTE, 
                 data = Summary_PriorTE, 
                 fun="event",
                 title = "TE Curves for Covariate Corrected Residuals",
                 ylab = "Probability of TE",
                 xlab = "Time in Years",
                 break.x.by = 0.5,
                 xlim=c(2.5,8),
                 conf.int = T,
                 pval = TRUE, pval.method = F,    ### Add p-value &  method name
                 surv.median.line = "hv",         ### Add median survival lines
                 legend.title = "Microbial cov",  ### Change legend titles
                 legend.labs = c("Old", "Young"), ### Change legend labels
                 palette = "nejm",         ### Use JCO journal color palette
                 risk.table = F,                  ### Add No at risk table
                 cumevents = F,                   ### Add cumulative No of events table
                 tables.height = 0.15,            ### Specify tables height
                 tables.theme = theme_cleantable(),### Clean theme for tables
                 tables.y.text = FALSE            ### Hide tables y axis text
                 )
p.coxph.cov$plot<-p.coxph.cov$plot+geom_vline(xintercept = 5.4, linetype = "solid", color="red")
p.coxph.cov
```

### Models {.tabset}
```{r}
coxph.TE.cov<-coxph(Surv(time,event)~resid_avg_TE
                    #+slope_covTE
                    #+age_avg_TE
                    +mompresent_mat
                   # +matsis_avg_TE
                    +lowMatRank
                    +excess_cfemales_avg_TE
                    +rain_avg_TE
                    +genome_wide_anubis_ancestry
                    ,data=Summary_PriorTE)
sink("model.txt")
summary(coxph.TE.cov);cox.zph(coxph.TE.cov)
sink()
ggforest(coxph.TE.cov,data=Summary_PriorTE,
                       main="TE Covariates Prior to Milestone")

```

# Dispersal {.tabset .tabset-pills}
## Prior to Dispersal {.tabset}
### Data
```{r}
PriorDisp<-males %>%  
  filter(case_when(!is.na(dispersed)&dispconfidence>=3&
                     dispersed>min(metadata_all$collection_date)~age_chrono<dispersed,
                   is.na(dispersed)&(status==0|status==2|status==3)~
                     age_chrono<max(age_dispersed,na.rm = T))) %>% 
  group_by(sname) %>% 
  mutate(rain_avg_Disp=mean(rain_yr,na.rm=T),
         matsis_avg_Disp=mean(matsisters_yr,na.rm=T),
         excess_cfemales_avg_Disp=mean(excess_cfemales_yr,na.rm=T),
         age_avg_Disp=mean(age_chrono)) %>% 
  ungroup() %>% drop_na(hydroyear)

PriorDisp$resid_cov<-resid(lmer(age_mb~age_chrono+ 
                                  avg_month_maxtemp+
                                  sum_month_rain+
                                  season+
                                  (1|collect_grp)+
                                  (1|hydroyear),
                                data=PriorDisp,
                                control=lmerControl(optCtrl=list(maxfun=2000))))

Summary_PriorDisp<-PriorDisp %>% 
  left_join(slope_M,by="sname") %>% 
  group_by(sname) %>% 
  mutate(resid_avg_Disp=mean(resid_cov,na.rm=T)) %>% 
  ungroup() %>% 
  group_by(sname,birth,statdate,resid_avg_Disp,dispersed,dispconfidence,age_dispersed,hybridscore,rain_avg_Disp,matsis_avg_Disp,excess_cfemales_avg_Disp,lowMatRank,mompresent_disp,age_avg_Disp,genome_wide_anubis_ancestry) %>% 
  summarise() %>% 
  ungroup() %>% 
  mutate(time=case_when(is.na(dispersed)~(statdate-birth)/365.25,
                       !is.na(dispersed)~(dispersed-birth)/365.25),
         event=case_when(is.na(dispersed)|(dispconfidence<3)~0,
                         !is.na(dispersed)&(dispconfidence>=3)~1),
         type_cov_accel=case_when(resid_avg_Disp>median(resid_avg_Disp)~"old",
                            resid_avg_Disp<median(resid_avg_Disp)~"young",
                            resid_avg_Disp==median(resid_avg_Disp)~"young")) 
#saveRDS(Summary_PriorDisp,"../out/DataForFigures/dispersal_prior_linear.rds")
```

### Plots
```{r}
survfit.cov.PriorDisp<-survfit(Surv(time,event)~type_cov_accel,
                          data=Summary_PriorDisp)
p.coxph.cov<-ggsurvplot(survfit.cov.PriorDisp, 
                 data = Summary_PriorDisp, 
                 fun="event",
                 title = "Dispersal Curves for Age Accel Prior to Milestone",
                 ylab = "Probability of Dispersal",
                 xlab = "Time in Years",
                 break.x.by = 1,
                 xlim=c(2.5,14),
                 conf.int = T,
                 pval = TRUE, pval.method = F,    ### Add p-value &  method name
                 surv.median.line = "hv",         ### Add median survival lines
                 legend.title = "Age Acceleration",  ### Change legend titles
                 legend.labs = c("Old", "Young"), ### Change legend labels
                 palette = "nejm",         ### Use JCO journal color palette
                 risk.table = F,                  ### Add No at risk table
                 cumevents = F,                   ### Add cumulative No of events table
                 tables.height = 0.15,            ### Specify tables height
                 tables.theme = theme_cleantable(),### Clean theme for tables
                 tables.y.text = FALSE            ### Hide tables y axis text
                 )
p.coxph.cov$plot<-p.coxph.cov$plot+geom_vline(xintercept = 7.47, linetype = "solid", color="red")
p.coxph.cov

```

### Models {.tabset .tabset-pills}
```{r}
coxph.Disp.cov<-coxph(Surv(time,event)~resid_avg_Disp
                     # +slope_covDisp
                    #  +age_avg_Disp
                      +mompresent_disp
                      +matsis_avg_Disp
                      +lowMatRank
                      #+excess_cfemales_avg_Disp
                      +rain_avg_Disp
                      #+genome_wide_anubis_ancestry
                      ,data=Summary_PriorDisp)
sink("model.txt")
summary(coxph.Disp.cov);cox.zph(coxph.Disp.cov)
sink()
ggforest(coxph.Disp.cov,data=Summary_PriorDisp,
         main="Dispersal Covariates Prior to Milestone")

```

# Male Rank Attainment {.tabset .tabset-pills}
## Prior to Rank Attainment {.tabset}
### Data
```{r}
PriorRank<-males %>%  
  filter(case_when(!is.na(ranked)&rstatus=="O"&
                     ranked>min(metadata_all$collection_date)~age_chrono<ranked,
                   is.na(ranked)&(status==0|status==2|status==3)~
                     age_chrono<max(metadata_all$age_ranked,na.rm = T))) %>% 
  group_by(sname) %>% 
  mutate(rain_avg_Rank=mean(rain_yr,na.rm=T),
         matsis_avg_Rank=mean(matsisters_yr,na.rm=T),
         excess_cfemales_avg_Rank=mean(excess_cfemales_yr,na.rm=T),
         age_avg_life=mean(age_chrono)) %>% 
  ungroup() %>% drop_na(hydroyear)
PriorRank$resid_cov<-resid(lmer(age_mb~age_chrono+ 
                                  avg_month_maxtemp+
                                  sum_month_rain+
                                  season+
                                  (1|collect_grp)+
                                  (1|hydroyear),
                                data=PriorRank,
                                control=lmerControl(optCtrl=list(maxfun=2000))))

Summary_PriorRank<-PriorRank %>% 
  group_by(sname) %>% 
  mutate(resid_avg_Rank=mean(resid_cov)) %>% 
  ungroup() %>% 
  group_by(sname,sex,birth,resid_avg_Rank,statdate,ranked,rstatus,age_ranked,hybridscore,rain_avg_Rank,excess_cfemales_avg_Rank,matsis_avg_Rank,mompresent_rank,lowMatRank,age_avg_life,genome_wide_anubis_ancestry) %>% 
  summarise() %>% 
  ungroup() %>% 
  mutate(time=case_when(is.na(ranked)~(statdate-birth)/365.25,
                       !is.na(ranked)~(ranked-birth)/365.25),
         event=case_when(is.na(ranked)~0,
                         !is.na(ranked)~1),
         type_cov_accel=case_when(resid_avg_Rank>median(resid_avg_Rank)~"old",
                            resid_avg_Rank<median(resid_avg_Rank)~"young",
                            resid_avg_Rank==median(resid_avg_Rank)~"old"))

#saveRDS(Summary_PriorRank,"../out/DataForFigures/rankattainment_males_prior_linear_2022.rds")
```

### Plots
```{r}
survfit.cov.PriorRank<-survfit(Surv(time,event)~type_cov_accel,
                          data=Summary_PriorRank)
p.coxph.cov<-ggsurvplot(survfit.cov.PriorRank, 
                 data = Summary_PriorRank, 
                 fun="event",
                 title = "Rank Attainment Curves for Age Accel Prior to Milestone",
                 ylab = "Probability of Rank Attainment",
                 xlab = "Time in Years",
                 break.x.by = 0.5,
                 xlim=c(4,10),
                 conf.int = T,
                 pval = TRUE, pval.method = F,    ### Add p-value &  method name
                 surv.median.line = "hv",         ### Add median survival lines
                 legend.title = "Microbial cov",  ### Change legend titles
                 legend.labs = c("Old", "Young"), ### Change legend labels
                 palette = "nejm",         ### Use JCO journal color palette
                 risk.table = F,                  ### Add No at risk table
                 cumevents = F,                   ### Add cumulative No of events table
                 tables.height = 0.15,            ### Specify tables height
                 tables.theme = theme_cleantable(),### Clean theme for tables
                 tables.y.text = FALSE            ### Hide tables y axis text
                 )
p.coxph.cov$plot<-p.coxph.cov$plot+geom_vline(xintercept = 7.38, linetype = "solid", color="red")
p.coxph.cov
```

### Models {.tabset .tabset-pills}
```{r}
coxph.Rank.cov<-coxph(Surv(time,event)~resid_avg_Rank
                     # +slope_covRank
                     # +age_avg_life
                      +mompresent_rank
                      +matsis_avg_Rank
                      +lowMatRank
                      +excess_cfemales_avg_Rank
                      +rain_avg_Rank
                      #+genome_wide_anubis_ancestry
                      ,data=Summary_PriorRank)
sink("model.txt")
summary(coxph.Rank.cov);cox.zph(coxph.Rank.cov)
sink()
ggforest(coxph.Rank.cov,data=Summary_PriorRank,
         main="Rank Covariates Prior to Milestone")
```

# Survival {.tabset .tabset-pills}
Age_death was calculated as the statdate-birth when status=1 and the statdate was between the start of sampling (~2000) and within a year of the end of sampling (~2015).
```{r}
SurvCov<-rbind(females,males) %>% 
  group_by(sname) %>% 
  mutate(resid_avg_life=mean(resid_life,na.rm=T)) %>% 
  ungroup() %>% 
  arrange(DADA_id) %>% 
  mutate(age_stat=as.numeric((statdate-birth)/365.25)) %>% 
  mutate(age_death=ifelse(status==1,age_stat,NA)) %>% 
  mutate(DeathAsJuv=case_when(age_death<=4~TRUE,
                              age_death>4|is.na(age_death)~FALSE))
                              #is.na(age_death)&age_stat<4~"CENSORED")) <- 0 cases of this
```

## Juvenile Survival {.tabset .tabset-pills}
### Females {.tabset}
#### Data
```{r}
JuvSurv<-SurvCov %>% 
  filter(age_chrono<=4&status<=3&sex=="F")

JuvSurv$resid_cov<-resid(lmer(age_mb~age_chrono
                              +avg_month_maxtemp
                              +sum_month_rain
                              +season
                              +(1|collect_grp)
                              +(1|hydroyear)
                              ,data=JuvSurv))

filter_JuvSurv<-JuvSurv %>% 
  group_by(sname) %>% 
  summarise(count=n()) %>% 
  ungroup() %>% 
  filter(count>=3)
JuvSurv<-JuvSurv %>% filter(sname %in% filter_JuvSurv$sname)

Summary_JuvSurv<-JuvSurv%>% 
  group_by(sname) %>% 
  mutate(resid_avg_life=mean(resid_cov,na.rm=T),
         age_avg_life=mean(age_chrono)) %>% 
  ungroup() %>% 
  mutate(time=case_when(DeathAsJuv=="FALSE"~4,
                        DeathAsJuv=="TRUE"~as.numeric((statdate-birth)/365.25)),
         event=case_when(is.na(age_death)~0,
                         !is.na(age_death)&DeathAsJuv=="FALSE"~0,
                         !is.na(age_death)&DeathAsJuv=="TRUE"~1),
         type_accel=case_when(resid_avg_life>median(resid_avg_life)~"old",
                             resid_avg_life<median(resid_avg_life)~"young",
                             resid_avg_life==median(resid_avg_life)~"young")) %>% 
  group_by(sname,sex,age_death,DeathAsJuv,resid_avg_life,maternal_loss,sibling,drought,highGrpSize,lowMatSCI,lowMatRank,cumulative_adversity,time,event,type_accel,type_pace,age_avg_life) %>% 
  summarise() %>% 
  ungroup()
```

#### Plots
```{r}
survfit.cov.lifespan<-survfit(Surv(time,event)~type_accel,
                          data=Summary_JuvSurv)
p.coxph.cov<-ggsurvplot(survfit.cov.lifespan, 
                 data = Summary_JuvSurv, 
                 #fun="event",
                 title = "Female Juv Survival Curves for Age Acceleration",
                 ylab = "Probability of Survival",
                 xlab = "Time in Years",
                 break.x.by = .5,
                 xlim=c(0,4),
                 conf.int = T,
                 pval = TRUE, pval.method = F,    ### Add p-value &  method name
                 surv.median.line = "hv",         ### Add median survival lines
                 legend.title = "Age Acceleration",  ### Change legend titles
                 legend.labs = c("Old", "Young"), ### Change legend labels
                 palette = "nejm",         ### Use JCO journal color palette
                 risk.table = F,                  ### Add No at risk table
                 cumevents = F,                   ### Add cumulative No of events table
                 tables.height = 0.15,            ### Specify tables height
                 tables.theme = theme_cleantable(),### Clean theme for tables
                 tables.y.text = FALSE            ### Hide tables y axis text
                 )
p.coxph.cov
```

#### Models {.tabset}
```{r}
coxphm.time.cov<-coxph(Surv(time, event)~resid_avg_life
                    #   +slope_prior4_1
                       #+age_avg_life
                       +maternal_loss
                       +sibling
                       +drought
                       +highGrpSize
                       +lowMatSCI
                       +lowMatRank
                       ,data=Summary_JuvSurv)
summary(coxphm.time.cov);cox.zph(coxphm.time.cov)
ggforest(coxphm.time.cov,data=Summary_JuvSurv,
         main="Juvenile Survival Covariates")
coxphm.time.cov2<-coxph(Surv(time, event)~resid_avg_life
                     #   +slope_prior4_1
                        #+age_avg_life
                        +cumulative_adversity
                        ,data=Summary_JuvSurv)
sink("model.txt")
summary(coxphm.time.cov2);cox.zph(coxphm.time.cov2)
sink()
ggforest(coxphm.time.cov2,data=Summary_JuvSurv,
         main="Juvenile Survival Covariates")
```

### Males {.tabset}
#### Data
```{r}
JuvSurv<-SurvCov %>% 
  filter(age_chrono<=4&status<=3&sex=="M") %>% drop_na(hydroyear)

JuvSurv$resid_cov<-resid(lmer(age_mb~age_chrono
                              +avg_month_maxtemp
                              +sum_month_rain
                              +season
                              +(1|collect_grp)
                              +(1|hydroyear)
                              ,data=JuvSurv))

Summary_JuvSurv<-JuvSurv%>% 
  group_by(sname) %>% 
  mutate(resid_avg_life=mean(resid_cov,na.rm=T),
         age_avg_life=mean(age_chrono)) %>% 
  ungroup() %>% 
  mutate(time=case_when(DeathAsJuv=="FALSE"~4,
                        DeathAsJuv=="TRUE"~as.numeric((statdate-birth)/365.25)),
         event=case_when(is.na(age_death)~0,
                         !is.na(age_death)&DeathAsJuv=="FALSE"~0,
                         !is.na(age_death)&DeathAsJuv=="TRUE"~1),
         type_accel=case_when(resid_avg_life>median(resid_avg_life)~"old",
                             resid_avg_life<median(resid_avg_life)~"young",
                             resid_avg_life==median(resid_avg_life)~"young"),
         type_pace=case_when(slope_prior4_1>median(slope_prior4_1)~"faster",
                               slope_prior4_1<median(slope_prior4_1)~"slower",
                               slope_prior4_1==median(slope_prior4_1)~"slower")) %>% 
  group_by(sname,sex,age_death,DeathAsJuv,resid_avg_life,maternal_loss,sibling,drought,highGrpSize,lowMatSCI,lowMatRank,cumulative_adversity,time,event,type_accel,type_pace,age_avg_life) %>% 
  summarise() %>% 
  ungroup()
```

#### Plots
```{r}
survfit.cov.lifespan<-survfit(Surv(time,event)~type_accel,
                          data=Summary_JuvSurv)
p.coxph.cov<-ggsurvplot(survfit.cov.lifespan, 
                 data = Summary_JuvSurv, 
                 #fun="event",
                 title = "Juv Survival Attainment Curves for Age Acceleration",
                 ylab = "Probability of Survival",
                 xlab = "Time in Years",
                 break.x.by = .5,
                 xlim=c(0,4),
                 conf.int = T,
                 pval = F, pval.method = F,    ### Add p-value &  method name
                 surv.median.line = "hv",         ### Add median survival lines
                 legend.title = "Age Acceleration",  ### Change legend titles
                 legend.labs = c("Old", "Young"), ### Change legend labels
                 palette = "nejm",         ### Use JCO journal color palette
                 risk.table = F,                  ### Add No at risk table
                 cumevents = F,                   ### Add cumulative No of events table
                 tables.height = 0.15,            ### Specify tables height
                 tables.theme = theme_cleantable(),### Clean theme for tables
                 tables.y.text = FALSE            ### Hide tables y axis text
                 )
p.coxph.cov
```

#### Models {.tabset}
```{r}
coxphm.time.cov<-coxph(Surv(time, event)~resid_avg_life
                      # +slope_prior4_1
                       +age_avg_life
                       +maternal_loss
                       +sibling
                       +drought
                       +highGrpSize
                       +lowMatSCI
                       +lowMatRank
                       ,data=Summary_JuvSurv)
summary(coxphm.time.cov);cox.zph(coxphm.time.cov)
ggforest(coxphm.time.cov,data=Summary_JuvSurv,
         main="Juvenile Survival Covariates")
coxphm.time.cov2<-coxph(Surv(time, event)~resid_avg_life
                       # +slope_prior4_1
                        +cumulative_adversity
                        ,data=Summary_JuvSurv)
sink("model.txt")
summary(coxphm.time.cov2);cox.zph(coxphm.time.cov2)
sink()
ggforest(coxphm.time.cov2,data=Summary_JuvSurv,
         main="Juvenile Survival Covariates")
```

## Adult Female Survival {.tabset .tabset-pills}
### Data
```{r}
SurvCovAdultF<-SurvCov %>% 
  filter(sex=="F",age_stat>4) #just removing those that died or disappeared before 4

filter_ADFSurv<-SurvCovAdultF %>% 
  group_by(sname) %>% 
  summarise(count=n()) %>% 
  ungroup() %>% 
  filter(count>=3)

ADFSurv<-SurvCovAdultF %>% filter(sname %in% filter_ADFSurv$sname)

Summary_ADFSurv<-ADFSurv %>% 
  group_by(sname) %>% 
  mutate(avg_agec=mean(age_chrono)) %>%
  ungroup() %>% 
  mutate(time=age_stat,
         event=case_when(is.na(age_death)~0,
                         !is.na(age_death)&DeathAsJuv=="FALSE"~1),
         type_cov_accel=case_when(resid_avg_life>median(resid_avg_life)~"old",
                            resid_avg_life<median(resid_avg_life)~"young",
                            resid_avg_life==median(resid_avg_life)~"young")) %>%
  group_by(sname,age_death,age_stat,resid_avg_life,time,event,type_cov_accel,DeathAsJuv,avg_agec)%>%
  summarise(rain_lifeavg=mean(rain_1yrprior,na.rm=T)) %>% 
  ungroup()

#saveRDS(Summary_ADFSurvCov,"../out/Summary_ADFSurvivalCovariates.rds")
```

### Plots
```{r}
survfit.cov.lifespan<-survfit(Surv(time,event)~type_cov_accel,
                          data=Summary_ADFSurvCov)
p.coxph.cov<-ggsurvplot(survfit.cov.lifespan, 
                 data = Summary_ADFSurvCov, 
                 #fun="event",
                 title = "ADF Survival Curves for Lifespan Age Accel",
                 ylab = "Probability of Survival",
                 xlab = "Time in Years",
                 break.x.by = 5,
                 xlim=c(4,30),
                 conf.int = T,
                 pval = TRUE, pval.method = F,    ### Add p-value &  method name
                 surv.median.line = "hv",         ### Add median survival lines
                 legend.title = "Age Acceleration",  ### Change legend titles
                 legend.labs = c("Old", "Young"), ### Change legend labels
                 palette = "nejm",         ### Use JCO journal color palette
                 risk.table = F,                  ### Add No at risk table
                 cumevents = F,                   ### Add cumulative No of events table
                 tables.height = 0.15,            ### Specify tables height
                 tables.theme = theme_cleantable(),### Clean theme for tables
                 tables.y.text = FALSE            ### Hide tables y axis text
                 )
p.coxph.cov
```

### Models
```{r}
coxph.Surv.cov<-coxph(Surv(time,event)~resid_avg_life
                      #+slope_life_1step
                      #+avg_agec
                      +maternal_loss
                      +sibling
                      +drought
                      +highGrpSize
                      +lowMatSCI
                      +lowMatRank
                      #+rain_lifeavg
                      +avg_SCI_F
                      +avg_SCI_M
                      +avg_proprank
                      ,data=Summary_ADFSurvCov)
summary(coxph.Surv.cov);cox.zph(coxph.Surv.cov)
ggforest(coxph.Surv.cov,data=Summary_ADFSurvCov,
         main="Adult Female Survival Covariates")
coxph.Surv.cov2<-coxph(Surv(time,event)~resid_avg_life
                      # +slope_life_1step
                       #+avg_agec
                       +cumulative_adversity
                       #+rain_lifeavg
                       +avg_SCI_F
                       +avg_SCI_M
                       +avg_proprank
                       ,data=Summary_ADFSurvCov)
summary(coxph.Surv.cov2);cox.zph(coxph.Surv.cov2)
ggforest(coxph.Surv.cov2,data=Summary_ADFSurvCov,
         main="Adult Female Survival Covariates") 

coxph.Surv.cov3<-coxph(Surv(time,event)~resid_avg_life
                      #+slope_over_age
                     # +avg_agec
                      +maternal_loss
                      +sibling
                      +drought
                      +highGrpSize
                      +lowMatSCI
                      +lowMatRank
                      #+rain_lifeavg
                      +avg_SCI_F
                      +avg_SCI_M
                      +avg_proprank
                      ,data=Summary_ADFSurvCov)
summary(coxph.Surv.cov3);cox.zph(coxph.Surv.cov3)
ggforest(coxph.Surv.cov3,data=Summary_ADFSurvCov,
         main="Adult Female Survival Covariates")
coxph.Surv.cov4<-coxph(Surv(time,event)~resid_avg_life
                      # +slope_over_age
                       #+avg_agec
                       +cumulative_adversity
                       #+rain_lifeavg
                       +avg_SCI_F
                       +avg_SCI_M
                       +avg_proprank
                       ,data=Summary_ADFSurvCov)
sink("model.txt")
summary(coxph.Surv.cov4);cox.zph(coxph.Surv.cov4)
sink()
ggforest(coxph.Surv.cov4,data=Summary_ADFSurvCov,
         main="Adult Female Survival Covariates") 
```
